requirements.txt
flask==2.3.3
pandas==2.1.1
numpy==1.24.3
plotly==5.17.0
yfinance==0.2.18
requests==2.31.0
python-dotenv==1.0.0
click==8.1.7